fix: File -> Save As breaks in some cases (#435)
In some cases, the NDMF preview scene could become the "active scene", which means
the File -> Save As menu item would attempt to save as that scene. This would then
result in the sceneSaving hook clearing the scene and an empty scene file being created.

Avoid this by always ensuring something other than the preview scene is selected;
also, close the preview scene (and create a replacement) if it's the last scene
left.