fix: freezes encountered by some users
Previously, `TaskUtil.OnMainThread` was asynchronous, allowing for the
editor to remain responsive even if something was spamming
invalidations. However, with the prior change fdaee911a3, this was
accidentally changed to be synchronous. This seems to cause editor
freezes in some situations - particularly, if OnMainThread was invoked
early in editor initialization, before we know which thread is the main
thread.

This change reverts this logic to be asynchronous.

NOTE: This also reverts the accidental addition of a public API in 1.6.3

Closes: #507