feat: add support for controlling platforms for pass execution
This adds new APIs to control which platforms a pass runs on.
For compatibility, if there is no platform support declaration, passes
will run only on VRCSDK 3.0.

Note that this only adds the APIs (and suppresses passes not compatible
with VRCSDK3.0). Full support will come in later changes, but this
change encompasses the semver minor release changes.