fix: curves with paths rewritten multiple times might be deleted (#598)
Previously, VirtualClip tried to track the modified state on a
curve-by-curve basis. This was meant to be an optimization to avoid
having to rewrite all curves when only some were modified, but
unfortunately due to weird Unity bugs, it turned out to be necessary to
always clear the clip and write all curves, making this tracking a bit
vestigial.

In some cases, the cached curves used to mark "deleted" curves would be
further rewritten if path rewriting happened multiple times, which could
end up overwriting (and deleting) a curve which should be retained.

This change removes this vestigial dirty tracking to remove this issue
and simplify the implementation overall.

Closes: https://github.com/bdunderscore/modular-avatar/issues/1550