fix: mesh preview fails on readonly meshes
Normally, Unity allows editor scripts to access readonly mesh data.
However, this specifically only applies to editor callback functions -
NDMF was running things in the Unity synchronization context, which was
not considered an editor callback. To work around this, we now perform
NDMFSyncContext processing in the editor update callback, which should
fix it for all preview plugins.