fix: compatibility break with parameter driver correction logic (#696)
* fix: compatibility break with parameter driver correction logic

The initial implementation of the parameter driver correction logic
would add new temporary parameters to the parameter driver, then add the
parameter to the controller. However, if the layer was present in
multiple controllers, it might not be corrected. In fact, this happens
in prior versions of MA, resulting in broken animator controllers.

To fix this, we defer the parameter creation to the Prepare stage just
before committing the controller.